refactor: renamed `price.rs` to `fee.rs`
