+++ omni-relayer/src/utils/mod.rs
@@ -1,4 +1,4 @@
+pub mod fee;
 pub mod near;
-pub mod price;
 pub mod redis;
 pub mod wormhole;
